+++ pcs/lib/commands/fencing_topology.py
@@ -1,7 +1,5 @@
 from typing import Optional
 
-from pcs.common import reports as report
-from pcs.common.fencing_topology import TARGET_TYPE_NODE
 from pcs.common.types import StringCollection
 from pcs.lib.cib import fencing_topology as cib_fencing_topology
 from pcs.lib.cib.tools import (
@@ -82,10 +80,6 @@ def remove_levels_by_params(
     target_type: Optional[str] = None,
     target_value=None,
     devices: Optional[StringCollection] = None,
-    # TODO remove, deprecated backward compatibility layer
-    ignore_if_missing: bool = False,
-    # TODO remove, deprecated backward compatibility layer
-    target_may_be_a_device: bool = False,
 ):
     """
     Remove specified fencing level(s).
@@ -95,8 +89,6 @@ def remove_levels_by_params(
     target_type -- the removed fencing level target value type
     mixed target_value -- the removed fencing level target value
     devices -- list of stonith devices of the removed fencing level
-    ignore_if_missing -- when True, do not report if level not found
-    target_may_be_a_device -- enables backward compatibility mode for old CLI
     """
     topology_el = get_fencing_topology(lib_env.get_cib())
     report_list = cib_fencing_topology.remove_levels_by_params(
@@ -105,64 +97,10 @@ def remove_levels_by_params(
         target_type,
         target_value,
         devices,
-        ignore_if_missing,
-        validate_device_ids=(not target_may_be_a_device),
     )
-
-    if not target_may_be_a_device or target_type != TARGET_TYPE_NODE:
-        if lib_env.report_processor.report_list(report_list).has_errors:
-            raise LibraryError()
-        lib_env.push_cib()
-        return
-
-    # TODO remove, deprecated backward compatibility mode
-    # CLI command parameters are: level, node, stonith, stonith... Both the
-    # node and the stonith list are optional. If the node is omitted and the
-    # stonith list is present, there is no way to figure it out, since there is
-    # no specification of what the parameter is. Hence the pre-lib code tried
-    # both. First it assumed the first parameter is a node. If that fence level
-    # didn't exist, it assumed the first parameter is a device. Since it was
-    # only possible to specify node as a target back then, this is enabled only
-    # in that case.
-    # CLI has no way to figure out what the first parameter is. Therefore, the
-    # lib must try both cases if asked to do so.
-    if not report.has_errors(report_list):
-        lib_env.report_processor.report_list(report_list)
-        lib_env.push_cib()
-        return
-
-    level_not_found = False
-    for report_item in report_list:
-        if (
-            report_item.message.code
-            == report.codes.CIB_FENCING_LEVEL_DOES_NOT_EXIST
-        ):
-            level_not_found = True
-            break
-    if not level_not_found:
-        lib_env.report_processor.report_list(report_list)
+    if lib_env.report_processor.report_list(report_list).has_errors:
         raise LibraryError()
-
-    target_and_devices = [target_value]
-    if devices:
-        target_and_devices.extend(devices)
-    report_list_second = cib_fencing_topology.remove_levels_by_params(
-        topology_el,
-        level,
-        None,
-        None,
-        target_and_devices,
-        ignore_if_missing,
-        validate_device_ids=(not target_may_be_a_device),
-    )
-    if not report.has_errors(report_list_second):
-        lib_env.report_processor.report_list(report_list_second)
-        lib_env.push_cib()
-        return
-
-    lib_env.report_processor.report_list(report_list)
-    lib_env.report_processor.report_list(report_list_second)
-    raise LibraryError()
+    lib_env.push_cib()
 
 
 def verify(lib_env: LibraryEnvironment):